Rename bit scalar package to bitwise
