+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/bitwise/BitFunctionUtils.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.function.scalar.bit;
+package org.apache.pinot.common.function.scalar.bitwise;
 
 import java.util.List;
 import org.apache.calcite.rel.type.RelDataType;
