+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/bitwise/PolymorphicBinaryIntegralScalarFunction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.function.scalar.bit;
+package org.apache.pinot.common.function.scalar.bitwise;
 
 import javax.annotation.Nullable;
 import org.apache.pinot.common.function.FunctionInfo;
