+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/bitwise/PolymorphicShiftScalarFunction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.function.scalar.bit;
+package org.apache.pinot.common.function.scalar.bitwise;
 
 import javax.annotation.Nullable;
 import org.apache.pinot.common.function.FunctionInfo;
@@ -26,23 +26,27 @@ import org.apache.pinot.common.utils.DataSchema.ColumnDataType;
 
 
 /**
- * Base class for polymorphic unary integral scalar functions.
+ * Base class for polymorphic integral shift scalar functions.
  *
  * <p>Implementations are stateless and thread-safe.
  */
-abstract class PolymorphicUnaryIntegralScalarFunction implements PinotScalarFunction {
+abstract class PolymorphicShiftScalarFunction implements PinotScalarFunction {
 
   @Nullable
   @Override
   public FunctionInfo getFunctionInfo(ColumnDataType[] argumentTypes) {
-    if (argumentTypes.length != 1) {
+    if (argumentTypes.length != 2) {
       return null;
     }
-    ColumnDataType argumentType = argumentTypes[0].getStoredType();
-    if (argumentType == ColumnDataType.INT) {
+    ColumnDataType valueType = argumentTypes[0].getStoredType();
+    ColumnDataType shiftType = argumentTypes[1].getStoredType();
+    if (!BitFunctionUtils.isIntegral(shiftType)) {
+      return null;
+    }
+    if (valueType == ColumnDataType.INT) {
       return intFunctionInfo();
     }
-    if (argumentType == ColumnDataType.LONG) {
+    if (valueType == ColumnDataType.LONG) {
       return longFunctionInfo();
     }
     return null;
@@ -51,12 +55,12 @@ abstract class PolymorphicUnaryIntegralScalarFunction implements PinotScalarFunc
   @Nullable
   @Override
   public FunctionInfo getFunctionInfo(int numArguments) {
-    return numArguments == 1 ? longFunctionInfo() : null;
+    return numArguments == 2 ? longFunctionInfo() : null;
   }
 
   @Override
   public PinotSqlFunction toPinotSqlFunction() {
-    return BitFunctionUtils.unaryIntegralSqlFunction(getName());
+    return BitFunctionUtils.shiftSqlFunction(getName());
   }
 
   protected abstract FunctionInfo intFunctionInfo();
